+++ CHANGELOG.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Added
 
 - Support for `#IF`, `#ELIF`, `#ELSE`
+- Auto completion of macros when assigning values
 
 ### Fixed